move WithTransact logic to basestore
